+++ CHANGES.rst
@@ -5,3 +5,7 @@ Changes
 ------------------
 
 - Add support for Python 3.
+- Eliminate ZODB dependency. Instead of raising a
+  ``ZODB.POSException.StorageError`` for unexpected ``TypeErrors``
+  during commit, the new class
+  ``nti.transactions.interfaces.CommitFailedError`` is raised.